backwards compatibility, refactor and node types
Added backwards compatibility with the API and move everything to the same file to simplify the understanding .
Also added a suggestion to distinguish between remote and local TPDO's and RPDO's.
