+++ README.rst
@@ -65,6 +65,13 @@ Quick start
 
 Here are some quick examples of what you can do:
 
+The PDOs can be access by to forms:
+
+**1st:** :code:`node.tpdo[n]` or :code:`node.rpdo[n]`
+
+**2nd:** :code:`node.pdo.tx[n]` or :code:`node.pdo.rx[n]`
+
+The :code:`n` is the PDO index (normaly 1 to 4). The second form of access is for backward compability.
 
 .. code-block:: python
 
@@ -115,7 +122,7 @@ Here are some quick examples of what you can do:
     node.nmt.state = 'OPERATIONAL'
 
     # Read a value from TPDO[1]
-    node.rpdo[1].wait_for_reception()
+    node.tpdo[1].wait_for_reception()
     speed = node.tpdo[1]['Velocity actual value'].phys
     val = node.tpdo['Some group.Some subindex'].raw
